[TASK] Rename old database fields which have a changed type
Some database fields have been changed with a different type. When you do a database compare, the fields will be changed to the new type, losing all its information. The update wizard is useless then.

This patch renames these field names so the old values stay for the update wizard.