[TASK] Add memoization for items and pages
The database can be queried multiple times for the same records within runtime. This is due to how the TCA works. TCA 'user' type and 'ItemProcFunc' calls are done at runtime, doing these same queries. Memoization is using a cache for storing the results of these queries, simply a runtime cache for object in this case